Fixed SearchBar logic
The search bar now searches for the SearchString as a substring in the NewsItem's title.

No news items are displayed if the SearchString matches none of the news title.

Pull-to-reload and pagination will be disabled when the searchbar is active. It will be reactivated when the searchbar Cancel button is clicked.

All loaded news are displayed when the search bar is empty or the Cancel button is clicked.